fix(config): parse CC_PREANALYSIS_DELAY_SECS as float, not int
.env.example ships CC_PREANALYSIS_DELAY_SECS=0.5 — and it should be
0.5, sub-second: services/pipeline/anthropic.py:1127 already casts it
to float at the point of use. config.py was still parsing the raw env
var with int(), so any install that migrated this key from
.env.example (i.e. every install going forward) crashed
langgraph-app on startup with
"ValueError: invalid literal for int() with base 10: '0.5'".

Also extends the previous commit's install.sh plausibility check with
a second, independent pass that cross-checks .env values against
config.py's own int(os.getenv(...)) call sites, rather than relying
solely on .env.example's committed value to signal the expected type
— that assumption is exactly what let this one slip through
undetected. Verified with a temporary copy of config.py that still
used int(): the check correctly flags the key for manual attention
instead of letting the container crash; with the fix applied it's
correctly silent.

+++ config.py
@@ -311,7 +311,7 @@ CC_HISTORY_COMPRESS_THRESHOLD  = int(os.getenv("CC_HISTORY_COMPRESS_THRESHOLD", 
 CC_HISTORY_COMPRESS_KEEP_TURNS = int(os.getenv("CC_HISTORY_COMPRESS_KEEP_TURNS", "8"))
 # Fallback delay (seconds) before the CC pre-analysis planner fires.
 # Overridden per-server via model_load_delay in Admin UI → Servers.
-CC_PREANALYSIS_DELAY_SECS = int(os.getenv("CC_PREANALYSIS_DELAY_SECS", "20"))
+CC_PREANALYSIS_DELAY_SECS = float(os.getenv("CC_PREANALYSIS_DELAY_SECS", "20"))
 
 # Master kill-switch for the infrastructure-side "1M+ context" path on CC tool
 # requests: ChromaDB context indexing (Tier-3), Tier-2 semantic-memory injection
